add second half of 7.12 -- blog removal now uses redux

+++ osa7/bloglist/frontend/src/reducers/blogsReducer.js
@@ -1,5 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit'
 import blogService from '../services/blogs'
+import { notifyWith } from './infoReducer'
 
 const initialState = []
 
@@ -13,11 +14,15 @@ const blogSlice = createSlice({
 
     appendBlog(state, action) {
       state.push(action.payload)
+    },
+
+    deleteBlog(state, action) {
+      return state.filter(b => b.id !== action.payload.id)
     }
   }
 })
 
-export const { setBlogs, appendBlog } = blogSlice.actions
+export const { setBlogs, appendBlog, deleteBlog } = blogSlice.actions
 
 export const initializeBlogs = () => {
   return async dispatch => {
@@ -29,14 +34,21 @@ export const initializeBlogs = () => {
 export const addBlog = blog => {
   return async dispatch => {
     const newBlog = await blogService.create(blog)
+    dispatch(notifyWith(`A new blog '${newBlog.title}' by '${newBlog.author}' added`))
     dispatch(appendBlog(newBlog))
   }
 }
 
 export const removeBlog = blog => {
   return async dispatch => {
-    await blogService.remove(blog.id)
-    dispatch(setBlogs(blogs.filter(b => b.id !== blog.id)))
+    const ok = window.confirm(
+      `Sure you want to remove '${blog.title}' by ${blog.author}`
+    )
+    if (ok) {
+      await blogService.remove(blog.id)
+      dispatch(notifyWith(`The blog' ${blog.title}' by '${blog.author} removed`))
+      dispatch(deleteBlog(blog))
+    }
   }
 }
